+++ ignite/templates/app/options.go
@@ -8,6 +8,8 @@ type Options struct {
 	BinaryNamePrefix string
 	ModulePath       string
 	AddressPrefix    string
+	// IncludePrefixes is used to filter the files to include from the generator
+	IncludePrefixes []string
 }
 
 // Validate that options are usable.